social_network/requirements.txt
asgiref==3.3.0
Django==3.1.3
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
requests==2.24.0
python-dotenv==0.15.0
